requirements.txt
python-telegram-bot==22.5
aiohttp==3.11.18
aiosqlite==0.21.0
websockets==12.0
python-dotenv==1.0.1
